This is an internal-only change mostly useful for debugging.  When the compiler is run internally we always use stdout.