@sbc100 2025-02-03
I use hyperfine to confirm that this change doesn't regress performance:

Before:

```
$ hyperfine './emcc hello.o'
Benchmark 1: ./emcc hello.o
  Time (mean ± σ):     456.8 ms ±   4.4 ms    [User: 382.4 ms, System: 169.2 ms]
  Range (min … max):   450.2 ms … 463.9 ms    10 runs
```

After:

```
$ hyperfine './emcc hello.o'
Benchmark 1: ./emcc hello.o
  Time (mean ± σ):     458.9 ms ±   3.5 ms    [User: 381.2 ms, System: 169.4 ms]
  Range (min … max):   453.7 ms … 464.0 ms    10 runs
```